ENG-37187 : Deserialize DNS messages carried in DNSTap record to Vector event
* Split dnstap source code
